==> shun_notebook/14_train_oversample.ipynb <==
Code taken from https://github.com/radekosmulski/whale/blob/master/oversample.ipynb

%matplotlib inline
import matplotlib.pyplot as plt
from fastai.vision import *
from fastai.metrics import accuracy
from fastai.basic_data import *
from skimage.util import montage

import pandas as pd
from torch import optim
import re

from utils import *

import fastai
from fastprogress import force_console_behavior
import fastprogress
fastprogress.fastprogress.NO_BAR = True
master_bar, progress_bar = force_console_behavior()
fastai.basic_train.master_bar, fastai.basic_train.progress_bar = master_bar, progress_bar

MODEL_PATH = "../model/"

df = pd.read_csv('../data/train.csv')
df.head()

fn2label = {row[1].Image: row[1].Id for row in df.iterrows()}
path2fn = lambda path: re.search('\w*\.jpg$', path).group(0)

val_fns = pd.read_pickle('../data/10_val_fns')

SZ = 224
BS = 64
NUM_WORKERS = 16
SEED=0

name = '14-res50-full-train'

df = pd.read_csv('../data/10_oversampled_train.csv')

data = (
    ImageItemList
        .from_df(df[df.Id != 'new_whale'], '../data/train-extracted-224', cols=['Image'])
        .split_by_valid_func(lambda path: path2fn(path) in val_fns)
        .label_from_func(lambda path: fn2label[path2fn(path)])
        .add_test(ImageItemList.from_folder('../data/test-extracted-224'))
        .transform(get_transforms(do_flip=False, max_zoom=1, max_warp=0, max_rotate=2), size=SZ, resize_method=ResizeMethod.SQUISH)
        .databunch(bs=BS, num_workers=NUM_WORKERS, path='../data')
        .normalize(imagenet_stats)
)

data

%%time

learn = create_cnn(data, models.resnet50, lin_ftrs=[2048], model_dir=MODEL_PATH)
learn.clip_grad();

%%time

learn.fit_one_cycle(24, 1e-2)
learn.save(f'{name}-stage-1')

learn.unfreeze()

max_lr = 1e-4
lrs = [max_lr/100, max_lr/10, max_lr]

%%time

learn.fit_one_cycle(24, lrs)
learn.save(f'{name}-stage-2')

SZ = 224 * 2
BS = 64 // 4
NUM_WORKERS = 16
SEED=0

data = (
    ImageItemList
        .from_df(df[df.Id != 'new_whale'], '../data/train-extracted-448', cols=['Image'])
        .split_by_valid_func(lambda path: path2fn(path) in val_fns)
        .label_from_func(lambda path: fn2label[path2fn(path)])
        .add_test(ImageItemList.from_folder('../data/test-extracted-448'))
        .transform(get_transforms(do_flip=False, max_zoom=1, max_warp=0, max_rotate=2), size=SZ, resize_method=ResizeMethod.SQUISH)
        .databunch(bs=BS, num_workers=NUM_WORKERS, path='../data')
        .normalize(imagenet_stats)
)

data

%%time

learn = create_cnn(data, models.resnet50, lin_ftrs=[2048], model_dir=MODEL_PATH)
learn.clip_grad();
learn.load(f'{name}-stage-2')
learn.freeze_to(-1)

%%time

learn.fit_one_cycle(10, 1e-2 / 4)
learn.save(f'{name}-stage-3-10')

%%time

learn.fit_one_cycle(3, 1e-2 / 4)
learn.save(f'{name}-stage-3-10-2')

%%time

learn.fit_one_cycle(14, 1e-2 / 4)
learn.save(f'{name}-stage-3-10-3')

%%time

learn.unfreeze()

max_lr = 1e-3 / 4
lrs = [max_lr/100, max_lr/10, max_lr]

learn.fit_one_cycle(22, lrs)
learn.save(f'{name}-stage-4')

learn = create_cnn(data, models.resnet50, lin_ftrs=[2048], model_dir=MODEL_PATH)
learn.clip_grad();
learn.load(f'13-res50-full-train-stage-3-10')
learn.freeze_to(-1)

# with oversampling
df = pd.read_csv('../data/10_oversampled_train_and_val.csv')

data = (
    ImageItemList
        .from_df(df, '../data/train-extracted-448', cols=['Image'])
        .split_by_valid_func(lambda path: path2fn(path) in val_fns)
        .label_from_func(lambda path: fn2label[path2fn(path)])
        .add_test(ImageItemList.from_folder('../data/test-extracted-448'))
        .transform(get_transforms(do_flip=False), size=SZ, resize_method=ResizeMethod.SQUISH)
        .databunch(bs=BS, num_workers=NUM_WORKERS, path='../data')
        .normalize(imagenet_stats)
)

data

%%time

learn = create_cnn(data, models.resnet50, lin_ftrs=[2048], model_dir=MODEL_PATH)
learn.clip_grad();
learn.load(f'13-res50-full-train-stage-3-10')
learn.freeze_to(-1)

%%time

learn.fit_one_cycle(2, 1e-2 / 4)
learn.save(f'{name}-stage-5')

%%time

learn.fit_one_cycle(4, 1e-2 / 4)
learn.save(f'{name}-stage-5-4')

name

learn.unfreeze()

max_lr = 1e-3 / 4
lrs = [max_lr/100, max_lr/10, max_lr]

learn.purge()

%%time

learn.fit_one_cycle(3, lrs)
learn.save(f'{name}-stage-6')

preds, _ = learn.get_preds(DatasetType.Test)

preds = torch.cat((preds, torch.ones_like(preds[:, :1])), 1)

preds[:, 5004] = 0.06

classes = learn.data.classes + ['new_whale']

create_submission(preds, learn.data, name, classes)

pd.read_csv(f'../subs/{name}.csv.gz').head()

pd.read_csv(f'../subs/{name}.csv.gz').Id.str.split().apply(lambda x: x[0] == 'new_whale').mean()

!kaggle competitions submit -c humpback-whale-identification -f ../subs/{name}.csv.gz -m "{name}"
