==> shun_notebook/17_resize_384.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt
from fastai.vision import *
from fastai.core import parallel
import re
from pathlib import Path

from utils import *

SZ = 384
NUM_WORKERS = 16

PATH = Path('../data')

(PATH/f'train-{SZ}').mkdir(exist_ok=True)
(PATH/f'test-{SZ}').mkdir(exist_ok=True)

def resize_img(path, _):
    PIL.Image.open(path).resize((SZ,SZ), resample=PIL.Image.BICUBIC).save((PATH/f'{path.parent.name}-{SZ}'/path.name))

%%time

files = list((PATH/'train').iterdir()) + list((PATH/'test').iterdir())
parallel(resize_img, files, NUM_WORKERS)

len(list((PATH/f'train-{SZ}').iterdir())), len(list((PATH/f'test-{SZ}').iterdir()))

len(list((PATH/f'train').iterdir())), len(list((PATH/f'test').iterdir()))
